Temporary fix for GCC search path
Specifying RUNTIME_PREFIX to /$TARGET/ caused intermediate xgcc to look
for libs in /$TARGET/lib during final gcc compilation. This is here
solved by setting RUNTIME_PREFIX and DEVEL_PREFIX to absolute path to
sdk tree.

In future it has to either be set to final SDK destination directory or
problem with ldscripts has to be solved in different way. Most likely
the former