Use automatic image and version from a tag
By the way, scheme for naming images and versions changes. Before it was
arch-compiler-library for image name and compiler-library versions for
image version, which followed old concept from early stages of a
project.

Now image name is simply the same as name of a branch. Tag in git adds
to that name a single number, starting with 1. Then git can append to
the description it generates number of commits since tag, commit id and
dirty indicator, if the tag is not matched exactly. All this becomes an
image version and becomes part of a tag in Docker in place of latest
given usually.