Fix magical bitness error
uClibc buildsystem complained about unknown bitness. The cause was lack
of _MIPS_SZPTR define. It most likely should be set by compiler. But as
we have ancient version of gcc, it may happen that it is broken.

Force it in uClibc config instead. There is a risk, it is a workaround
for the particular error, not a fix for whole system.

Gcc-Before: 4.6.4
Gcc-After: 3.3.2